hack in some deleting from the page

+++ server/routes/api.js
@@ -227,11 +227,18 @@ router.post('/task', createOrEditTask);
 
 router.delete('/task', function(req, res, next){	
 	using(db.getConnection(), function(conn){
-		return index.deleteTask(conn, req.query.task_id);
+		return new Promise(function(resolve, reject){
+			if(index.isTaskCreator(conn, req.query.task_id, req.query.user_id)) {
+				reject('Not the task creator');
+			} else {
+				return index.deleteTask(conn, req.query.task_id);
+			}
+		});
 	}).then(function(){
+		log('Task ' + req.query.task_id + ' deleted by user ' + req.query.user_id);
 		res.json({success: true});
 	}).catch(function(err){
-		next(new ApiError(err, 'Failed to delete task and its relationships'))
+		next(new ApiError(err, 'Failed to delete task ' + req.query.task_id + ' and its relationships by user ' + req.query.user_id))
 	});
 });
